New Class Structure
The GEMPAK reader class structure has been updated and separates out
each type of file (grid, sounding, surface). This should make things
a little cleaner. The data structures within the class have also been
changed and now variable names should be properly converted for much
easier use.

The biggest change with this commit is that grids can now be read
in and unpacked, at least for certain packing types. For now, xarray
DataArray strucutres will be used to contain the grids. Merging into
a Dataset does not work as expected yet.

Work needs to be done to figure out why certain grids with the same
packing type are not processed correctly. It seems to be a bit order
issue that is not handled correctly by Python.