Fixed bit shifting issue
Figured out bit shifting problem with negative numbers in Python.
It turns out they have infinite bits at and to the left of the bit
inidcating pos/neg. A bit mask was initially applied at the wrong
step in the calculation. The bit mask to reduce it to a 4 byte int
needs to be applied to the raw packed data before shifting for it
to produce the correct result.