Restructure package
Several changes. Removed MetPy dependency as only a couple of functions
are needed. This requires using ported GEMPAK functions. Xarray is still
a dependency. There will need to be more refactoring in order to output
proper xarray objects as well as be able to search for specific data
within the GEMPAK files. Always reading the full file is not typically
what is wanted.