Finished surface data; fixed sounding bugs
Surface data can now read in and is complete barring any bugs.

Found bugs in the sounding merging code. Some due to errors in the
parameter calculations, others in the loops through the binary data.
Those have been fixed and data still appears reasonable.

Chnaged how data is read and output. Instead of just taking a whole
file, now you are able to select based on parameters. This makes things
quicker as you typically are not asking for the full file. The full file
will still be read if you do not use any parameters. xarray objects are
used for the grids and soundings, but JSON is used for surface data.