Added examples; fixed bugs
An notebook with examples and accompanying data were added.

Fixed an issue with data source numbers that were not part of the Enum
that defined the sources. Any value that cannot be interpreted now will
become an Unknown.

Fixed an issue with certain parameters not being decoded from bytes to
charcter. This caused problems when the bytes were used as dict keys
later.

Fixed an issue with surface data scalars being returned as arrays.

Added better detail to a few docstrings.