Improved performance and surface data processing
Contains decent performance improvements when reading from files. This
was done by eliminating the need to loop through every data header
multiple times to determine what to extract. The pointers to the data
are now determined once and then only those are used to extract data.

Surface data processing was improved so that the nearest observations
would be extracted. Special observations are now partially decoded so
that they can be also be returned. More information can be found at
Unidata/MetPy#2759.