Fix for improper data location
Column and row headers were handled in such a way that the wrong data
could be written to the file for a given row/column pair. This occurred
due to those headers being stored within sets. At various points, the
sets would be sorted into lists, but the order could be different than
expected during the actual writing step as they were not sorted. The
sets remain for removing duplicates, but now they are always sorted and
placed into lists as grids/parameters are added. Order is retained
properly during the write step and the issue should be fixed.