Fix grid encoding (#12)
Need to account for differences in the way times are encoded into
GEMPAK files based on whether they are an analysis or forecast.

Fix issue with analysis grids not having the right time written into the
GEMPAK grid. Bumped version.

Fixed issue when user inputs only a grid type to the date_time string
but not a forecast time. Time will now default to zero.

Removed TVM/UTM as they appear to not be supported since an old GEMPAK
version. Fixed issues with LEA/GNO projections as pyproj (or PROJ?) does
not seem to convert to CF format. Added several additional tests for
reading and writing all grid projections.